поправил апи статистики

+++ web/src/main/java/net/danil/web/statistics/dto/SolvedAttemptedCounts.java
@@ -1,7 +1,7 @@
 package net.danil.web.statistics.dto;
 
 public interface SolvedAttemptedCounts {
-    Long getSolvedCount();
+    Integer getSolvedCount();
 
-    Long getAttemptedCount();
+    Integer getAttemptedCount();
 }
